refactor(viewer): split AuditLogTab into components/admin/auditLog
1,732-line tab becomes a folder: a thin AuditLogTab on AdminTabShell +
DataTable, columns.tsx (desktop columns, WASM/issue-bot badges),
AuditEntryCards (mobile), MetricsControls, useAuditLogEntries (keyset
pagination), and the details modal split into DetailsModal, MetricsTab,
ClientMetricsTab, MetricsHistoryChart and ProfileStatsTable (now sorted
by DataTable). Markup, class names and strings unchanged.
