--- src/frontend/src/components/admin/AuditLogTab.tsx
@@ -1,1732 +1,0 @@
-import React, {useEffect, useMemo, useState} from "react";
-
-import {useAuditFilterStore} from "@/state/auditFilterStore";
-import {localDateTime} from "@/utils/time";
-import {
-    ApiError,
-    AuditEntry,
-    AuditFilters,
-    MetricsSample,
-    ProfileStatsRow,
-    viewerApi,
-} from "@/services/viewerApi";
-import {isMissingManifest, MISSING_MANIFEST_NOTE} from "./workerPackages";
-import {formatBytes as formatBytesBase, formatMillis} from "@/utils/format";
-
-// Filterable audit log view. Two layouts:
-// * sm:↑ desktop — table with sticky header, fits everything in columns.
-// * mobile — collapsible filters + card-per-entry, so a 320px viewport
-//   stays readable without horizontal scrolling.
-//
-// Pagination is keyset on the BIGSERIAL id (the server returns
-// next_before_id) — that way the table doesn't shift while new audit
-// rows are inserted between pages.
-
-// "compile" = one procedural compile RUN (app.py _audit_compile_run); its
-// worker attaches the run's log_key, so such a row's Log tab shows the engine
-// output for exactly that run.
-
-const PROFILE_SETTING_KEY = "profile_conversions";
-
-// Whether a row is a candidate for the (i) details modal. Convert
-// rows always qualify (they may have metrics, traceback, or both);
-// other rows only show the icon when they actually have an error or
-// traceback to surface, so the column doesn't fill with no-op
-// buttons.
-function hasDetails(e: AuditEntry): boolean {
-    if (e.action === "convert") return true;
-    // A compile run always has a log to read, even when it succeeded.
-    if (e.action === "compile") return true;
-    // A plugin job records the same metrics a convert does -- the worker passes
-    // them to _audit_done on every outcome -- so a SUCCESSFUL one has something
-    // worth opening. Without this it fell through to the error-only rule below
-    // and offered no icon at all, leaving those metrics collected but
-    // unreachable.
-    if (e.action === "plugin_job") return true;
-    // Browser view/render rows always carry a client_metrics payload to
-    // inspect (per-phase split + per-function frames), fetched lazily.
-    if (e.action === "view" || e.action === "render") return true;
-    return Boolean(e.error || e.traceback);
-}
-
-function isClientMetricsRow(e: AuditEntry): boolean {
-    return e.action === "view" || e.action === "render";
-}
-
-function hasMetrics(e: AuditEntry): boolean {
-    return (
-        e.cpu_user_ms != null ||
-        e.cpu_sys_ms != null ||
-        e.peak_rss_kb != null ||
-        e.read_bytes != null ||
-        e.write_bytes != null ||
-        e.profile_key != null ||
-        e.duration_ms != null
-    );
-}
-
-const AuditLogTab: React.FC = () => {
-    // The filter is owned by the Audit tab, not by this sub-tab: Overview
-    // counts the same population and the operator drills from one into the
-    // other. See state/auditFilterStore.
-    const filters = useAuditFilterStore((st) => st.filters);
-    const refreshNonce = useAuditFilterStore((st) => st.refreshNonce);
-    const [entries, setEntries] = useState<AuditEntry[]>([]);
-    const [nextBeforeId, setNextBeforeId] = useState<number | null>(null);
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState<string | null>(null);
-    const [detailsEntry, setDetailsEntry] = useState<AuditEntry | null>(null);
-    const [profileEnabled, setProfileEnabled] = useState(false);
-    const [profileSaving, setProfileSaving] = useState(false);
-    const [clearing, setClearing] = useState(false);
-
-    // Initial fetch of the profile-conversions toggle. Failures are
-    // non-fatal — the row still renders, the toggle just stays off.
-    useEffect(() => {
-        let cancelled = false;
-        (async () => {
-            try {
-                const v = await viewerApi.adminGetSetting(PROFILE_SETTING_KEY);
-                if (!cancelled) setProfileEnabled((v || "").toLowerCase() === "true");
-            } catch {
-                /* ignore */
-            }
-        })();
-        return () => {
-            cancelled = true;
-        };
-    }, []);
-
-    const onProfileToggle = async (next: boolean) => {
-        setProfileSaving(true);
-        try {
-            await viewerApi.adminSetSetting(PROFILE_SETTING_KEY, next ? "true" : "false");
-            setProfileEnabled(next);
-        } catch (e) {
-            setError(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setProfileSaving(false);
-        }
-    };
-
-    const onClearMetrics = async () => {
-        if (!window.confirm(
-            "Clear all conversion metrics and delete profile blobs? Audit rows themselves stay; only the metrics columns are nulled."
-        )) return;
-        setClearing(true);
-        try {
-            const r = await viewerApi.adminClearMetrics();
-            await reload(filters);
-            window.alert(
-                `Cleared ${r.rows_cleared} row(s); deleted ${r.profiles_deleted} profile blob(s).` +
-                (r.errors.length ? `\n${r.errors.length} error(s) — see browser console.` : "")
-            );
-            if (r.errors.length) console.warn("clear metrics errors", r.errors);
-        } catch (e) {
-            setError(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setClearing(false);
-        }
-    };
-
-    const reload = async (f: AuditFilters) => {
-        setLoading(true);
-        setError(null);
-        try {
-            const r = await viewerApi.adminAudit({...f, before_id: undefined});
-            setEntries(r.entries);
-            setNextBeforeId(r.next_before_id);
-        } catch (e) {
-            setError(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const loadMore = async () => {
-        if (nextBeforeId == null) return;
-        setLoading(true);
-        try {
-            const r = await viewerApi.adminAudit({...filters, before_id: nextBeforeId});
-            setEntries((prev) => [...prev, ...r.entries]);
-            setNextBeforeId(r.next_before_id);
-        } catch (e) {
-            setError(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    // Reacts to the shared filter rather than owning it: the bar lives in
-    // AuditTab, and Overview's tiles write to the same store, so a drill-down
-    // arrives here as a filter change like any other.
-    useEffect(() => {
-        void reload(filters);
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [filters, refreshNonce]);
-
-
-    return (
-        <div className="flex flex-col h-full">
-            <div className="border-b border-gray-700">
-                {/* Per-deployment knobs that affect future runs.
-                    Profile toggle persists in app_settings; Clear
-                    metrics nulls out columns + deletes blobs.
-                    Visually separated so the controls are obvious on
-                    mobile (where they otherwise sit flush with the
-                    Filters / Refresh row). */}
-                <div className="flex flex-wrap items-center gap-3 px-3 sm:px-4 py-2 text-xs border-t border-gray-800 bg-gray-900/40">
-                    <span className="font-semibold text-gray-300 uppercase tracking-wide text-[10px]">
-                        Metrics
-                    </span>
-                    <label className="flex items-center gap-2 cursor-pointer">
-                        <input
-                            type="checkbox"
-                            checked={profileEnabled}
-                            onChange={(e) => onProfileToggle(e.target.checked)}
-                            disabled={profileSaving}
-                            className="h-4 w-4"
-                        />
-                        <span>
-                            Profile conversions
-                            {profileSaving ? <span className="text-gray-400"> (saving…)</span> : null}
-                        </span>
-                    </label>
-                    <button
-                        className="ml-auto bg-gray-700 hover:bg-gray-600 px-2 py-1 rounded-sm disabled:opacity-50"
-                        onClick={onClearMetrics}
-                        disabled={clearing}
-                        title="Null out all metrics columns and delete profile blobs"
-                    >
-                        {clearing ? "Clearing…" : "Clear metrics"}
-                    </button>
-                </div>
-            </div>
-            {error && (
-                <div className="px-3 sm:px-4 py-2 text-red-300 text-xs border-b border-gray-700">
-                    {error}
-                </div>
-            )}
-            <div className="flex-1 min-h-0 overflow-auto">
-                {/* Desktop / tablet table.
-                    Time/User/Scope/Action/Target/Status are predictable in
-                    width, so we fix them and let Key flex to consume the
-                    rest of the row. Removes the postage-stamp truncation
-                    that kicked in even when the modal had room to show
-                    the full path. */}
-                {/* ``table-auto`` instead of ``table-fixed`` so each
-                    column sizes to its content rather than the
-                    rigid colgroup widths. ``min-w`` on each col
-                    pins a floor so a stale w-16 doesn't truncate a
-                    5-digit ID. ``whitespace-nowrap`` on the Td
-                    component below ensures content doesn't wrap to
-                    a second line when the natural width exceeds the
-                    floor. Key column stays flexible (``min-w-[14rem]``)
-                    so long source paths get the leftover width. */}
-                <table className="hidden sm:table w-full text-sm min-w-[1260px]">
-                    <colgroup>
-                        <col className="min-w-[5rem]"/>
-                        <col className="min-w-[11rem]"/>
-                        <col className="min-w-[8rem]"/>
-                        <col className="min-w-[12rem]"/>
-                        <col className="min-w-[6rem]"/>
-                        <col className="min-w-[14rem]"/>
-                        <col className="min-w-[6rem]"/>
-                        <col className="min-w-[6rem]"/>
-                    </colgroup>
-                    <thead className="sticky top-0 bg-gray-800">
-                    <tr className="text-left">
-                        <Th>ID</Th>
-                        <Th>Time</Th>
-                        <Th>User</Th>
-                        <Th>Scope</Th>
-                        <Th>Action</Th>
-                        <Th>Device</Th>
-                        <Th>Key</Th>
-                        <Th>Target</Th>
-                        <Th>Status</Th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    {entries.map((e) => (
-                        <tr key={e.id} className="border-t border-gray-800 hover:bg-gray-800/40">
-                            <Td>
-                                <span className="font-mono text-gray-300">#{e.id}</span>
-                            </Td>
-                            <Td title={e.ts || ""}>{formatTs(e.ts)}</Td>
-                            <Td title={userTooltip(e)}>{e.user_display_name || shortSub(e.user_sub)}</Td>
-                            <Td title={e.scope_id || ""}>
-                                {e.scope_kind}
-                                {e.scope_id ? `:${shortSub(e.scope_id)}` : ""}
-                            </Td>
-                            <Td>{e.action}</Td>
-                            <Td title={e.device_id || ""}>
-                                {e.device_id ? (
-                                    <span className="font-mono text-gray-400">{e.device_id.slice(0, 8)}</span>
-                                ) : (
-                                    <span className="text-gray-600">—</span>
-                                )}
-                            </Td>
-                            <Td title={e.key || ""}>{e.key || ""}</Td>
-                            <Td>
-                                {e.target_format || ""}
-                                {isWasmEntry(e) && <WasmBadge/>}
-                            </Td>
-                            <Td title={e.error || ""}>
-                                <span className={statusClass(e.status)}>{e.status || ""}</span>
-                                {hasDetails(e) && (
-                                    <button
-                                        type="button"
-                                        className="ml-1 inline-flex items-center justify-center w-4 h-4 rounded-full border border-gray-500 text-gray-300 hover:text-white hover:border-white text-[10px] font-bold leading-none align-middle no-drag"
-                                        onClick={() => setDetailsEntry(e)}
-                                        title={e.error ? "Show error / metrics" : "Show metrics"}
-                                        aria-label="Show details"
-                                    >
-                                        i
-                                    </button>
-                                )}
-                                <IssueBotBadge entry={e} onChanged={() => reload(filters)}/>
-                            </Td>
-                        </tr>
-                    ))}
-                    </tbody>
-                </table>
-                {/* Mobile cards */}
-                <ul className="sm:hidden divide-y divide-gray-800">
-                    {entries.map((e) => (
-                        <li key={e.id} className="px-3 py-2 text-xs">
-                            <div className="flex items-baseline justify-between gap-2">
-                                <span className="font-medium">
-                                    <span className="font-mono text-gray-400 mr-1">#{e.id}</span>
-                                    {e.action}
-                                </span>
-                                <span className={statusClass(e.status) + " text-[11px]"}>
-                                    {e.status || ""}
-                                </span>
-                            </div>
-                            <div className="text-gray-400 mt-0.5">
-                                {formatTs(e.ts)} · {e.scope_kind}
-                                {e.scope_id ? `:${shortSub(e.scope_id)}` : ""}
-                            </div>
-                            {e.key && (
-                                <div className="text-gray-300 mt-1 break-all" title={e.key}>
-                                    {e.key}
-                                    {e.target_format ? (
-                                        <span className="text-gray-400"> → {e.target_format}</span>
-                                    ) : null}
-                                    {isWasmEntry(e) && <WasmBadge/>}
-                                </div>
-                            )}
-                            {e.user_sub && (
-                                <div className="text-gray-500 mt-0.5" title={userTooltip(e)}>
-                                    by {e.user_display_name || shortSub(e.user_sub)}
-                                </div>
-                            )}
-                            {e.error && (
-                                <div className="text-red-300 mt-1 break-all flex items-start gap-1" title={e.error}>
-                                    <span className="flex-1">{e.error}</span>
-                                    <button
-                                        type="button"
-                                        className="shrink-0 inline-flex items-center justify-center w-4 h-4 rounded-full border border-gray-500 text-gray-300 hover:text-white hover:border-white text-[10px] font-bold leading-none mt-0.5 no-drag"
-                                        onClick={() => setDetailsEntry(e)}
-                                        aria-label="Show details"
-                                    >
-                                        i
-                                    </button>
-                                </div>
-                            )}
-                            {!e.error && hasDetails(e) && (
-                                <div className="mt-1">
-                                    <button
-                                        type="button"
-                                        className="text-[10px] text-gray-400 hover:text-white underline no-drag"
-                                        onClick={() => setDetailsEntry(e)}
-                                    >
-                                        details
-                                    </button>
-                                </div>
-                            )}
-                        </li>
-                    ))}
-                </ul>
-                {!loading && entries.length === 0 && (
-                    <div className="px-4 py-8 text-center text-gray-500 text-sm">
-                        No matching audit entries.
-                    </div>
-                )}
-            </div>
-            <div className="border-t border-gray-700 px-3 sm:px-4 py-2 flex items-center gap-3 text-xs">
-                <span className="text-gray-400">{entries.length} rows</span>
-                <button
-                    className="bg-gray-700 hover:bg-gray-600 px-3 py-1 rounded-sm disabled:opacity-50"
-                    onClick={loadMore}
-                    disabled={loading || nextBeforeId == null}
-                >
-                    Load more
-                </button>
-                {loading && <span className="text-gray-500">loading…</span>}
-            </div>
-            {detailsEntry && (
-                <DetailsModal entry={detailsEntry} onClose={() => setDetailsEntry(null)}/>
-            )}
-        </div>
-    );
-};
-
-// Lazy-loaded conversion log (worker-captured stdout/stderr, stored at log_key). Fetched on first
-// view of the Log tab — surfaces what the structured fields don't, e.g. an "adacpp-native ...
-// falling back to libtess2" warning when a pipeline silently degrades inside the convert subprocess.
-const LogTab: React.FC<{entry: AuditEntry}> = ({entry}) => {
-    const [text, setText] = useState<string | null>(null);
-    const [loading, setLoading] = useState(true);
-    const [err, setErr] = useState<string | null>(null);
-
-    useEffect(() => {
-        let alive = true;
-        setLoading(true);
-        setErr(null);
-        viewerApi
-            .adminGetAuditLog(entry.id)
-            .then((t) => {
-                if (alive) setText(t);
-            })
-            .catch((e) => {
-                if (!alive) return;
-                setErr(
-                    e instanceof ApiError
-                        ? e.status === 404
-                            ? "No log captured for this row."
-                            : e.detail || e.message
-                        : String(e),
-                );
-            })
-            .finally(() => {
-                if (alive) setLoading(false);
-            });
-        return () => {
-            alive = false;
-        };
-    }, [entry.id]);
-
-    if (loading) return <div className="p-4 text-sm text-gray-400">Loading log…</div>;
-    if (err) return <div className="p-4 text-sm text-gray-400 italic">{err}</div>;
-    return (
-        <pre className="p-3 text-xs leading-relaxed whitespace-pre-wrap break-words text-gray-200 font-mono">
-            {text}
-        </pre>
-    );
-};
-
-// Tabbed details view: Error (or 'OK' summary) on one tab, Metrics
-// on the other. Both tabs render even when their data is partial so
-// the user gets a consistent layout regardless of job outcome — a
-// timed-out conversion and a clean success share the same shape.
-const DetailsModal: React.FC<{entry: AuditEntry; onClose: () => void}> = ({entry, onClose}) => {
-    const clientRow = isClientMetricsRow(entry);
-    const [tab, setTab] = useState<"error" | "metrics" | "client" | "log">(
-        entry.error || entry.traceback ? "error" : clientRow ? "client" : "metrics",
-    );
-    const [copied, setCopied] = useState(false);
-    const [downloading, setDownloading] = useState(false);
-    const [downloadErr, setDownloadErr] = useState<string | null>(null);
-
-    useEffect(() => {
-        const onKey = (e: KeyboardEvent) => {
-            if (e.key === "Escape") onClose();
-        };
-        window.addEventListener("keydown", onKey);
-        return () => window.removeEventListener("keydown", onKey);
-    }, [onClose]);
-
-    const traceText = entry.traceback || entry.error || "";
-    const copyPayload =
-        (entry.error ? `${entry.error}\n\n` : "") + (entry.traceback || "");
-
-    const onCopy = async () => {
-        try {
-            await navigator.clipboard.writeText(copyPayload || traceText);
-            setCopied(true);
-            setTimeout(() => setCopied(false), 1500);
-        } catch {
-            /* clipboard blocked — user can still select-and-copy by hand */
-        }
-    };
-
-    const onDownloadProfile = async () => {
-        if (!entry.profile_key) return;
-        setDownloading(true);
-        setDownloadErr(null);
-        try {
-            // Suggest the storage filename so the user gets a stable
-            // name on disk; tail of the key after the last slash.
-            const suggested = entry.profile_key.split("/").pop() || `audit-${entry.id}.prof`;
-            await viewerApi.adminDownloadProfile(entry.id, suggested);
-        } catch (e) {
-            setDownloadErr(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setDownloading(false);
-        }
-    };
-
-    return (
-        <div
-            className="fixed inset-0 z-60 flex items-start sm:items-center justify-center bg-black/70 p-4 overflow-y-auto"
-            onClick={onClose}
-        >
-            <div
-                className="bg-gray-900 border border-gray-700 rounded-sm shadow-xl flex flex-col max-w-3xl w-full max-h-[calc(100dvh-2rem)] sm:max-h-[85dvh] my-auto"
-                onClick={(e) => e.stopPropagation()}
-                role="dialog"
-                aria-label="Audit row details"
-            >
-                <div className="flex items-start gap-3 border-b border-gray-700 px-4 py-2">
-                    <div className="flex-1 min-w-0">
-                        <div className="text-sm font-semibold">Audit details</div>
-                        <div className="text-xs text-gray-400 truncate" title={entry.key || ""}>
-                            {entry.action}
-                            {entry.key ? ` · ${entry.key}` : ""}
-                            {entry.target_format ? ` → ${entry.target_format}` : ""}
-                        </div>
-                    </div>
-                    {/* In the HEADER, not on a tab. It is an action on the job,
-                        not part of any one view of it — and a tab called
-                        "Outcome" is the last place someone looks for a button
-                        that changes the outcome. Renders nothing unless the row
-                        is actually cancellable, so it costs no space on the
-                        overwhelming majority of rows, which are history. */}
-                    <StuckJobActions entry={entry}/>
-                    {tab === "error" && entry.traceback && (
-                        <button
-                            type="button"
-                            className="shrink-0 bg-gray-800 hover:bg-gray-700 text-gray-100 px-2 py-1 rounded-sm text-xs"
-                            onClick={onCopy}
-                            title="Copy traceback to clipboard"
-                        >
-                            {copied ? "Copied" : "Copy"}
-                        </button>
-                    )}
-                    <button
-                        type="button"
-                        className="shrink-0 text-gray-300 hover:text-white text-xl leading-none px-2"
-                        onClick={onClose}
-                        aria-label="Close"
-                        title="Close (Esc)"
-                    >
-                        ×
-                    </button>
-                </div>
-                <div className="flex border-b border-gray-700 text-xs">
-                    {clientRow && (
-                        <TabButton active={tab === "client"} onClick={() => setTab("client")}>
-                            {entry.action === "render" ? "Render" : "Load"}
-                        </TabButton>
-                    )}
-                    <TabButton active={tab === "error"} onClick={() => setTab("error")}>
-                        {entry.error ? "Error" : "Outcome"}
-                    </TabButton>
-                    {!clientRow && (
-                        <TabButton active={tab === "metrics"} onClick={() => setTab("metrics")}>
-                            Metrics
-                            {hasMetrics(entry) ? null : (
-                                <span className="text-gray-500"> (none)</span>
-                            )}
-                        </TabButton>
-                    )}
-                    {!clientRow && (
-                        <TabButton active={tab === "log"} onClick={() => setTab("log")}>
-                            Log
-                        </TabButton>
-                    )}
-                </div>
-                <div className="flex-1 overflow-auto">
-                    {tab === "error" && (
-                        <ErrorTab entry={entry}/>
-                    )}
-                    {tab === "metrics" && (
-                        <MetricsTab
-                            entry={entry}
-                            onDownloadProfile={onDownloadProfile}
-                            downloading={downloading}
-                            downloadErr={downloadErr}
-                        />
-                    )}
-                    {tab === "client" && <ClientMetricsTab entry={entry}/>}
-                    {tab === "log" && <LogTab entry={entry}/>}
-                </div>
-            </div>
-        </div>
-    );
-};
-
-const TabButton: React.FC<{active: boolean; onClick: () => void; children: React.ReactNode}> = ({
-    active, onClick, children,
-}) => (
-    <button
-        type="button"
-        onClick={onClick}
-        className={
-            "px-3 py-1.5 border-b-2 " +
-            (active
-                ? "border-blue-500 text-white"
-                : "border-transparent text-gray-400 hover:text-gray-200")
-        }
-    >
-        {children}
-    </button>
-);
-
-const ErrorTab: React.FC<{entry: AuditEntry}> = ({entry}) => {
-    if (entry.error || entry.traceback) {
-        return (
-            <>
-                {entry.error && (
-                    <div className="px-4 py-2 text-sm text-red-300 border-b border-gray-800 wrap-break-word">
-                        {entry.error}
-                    </div>
-                )}
-                {entry.traceback ? (
-                    <pre className="px-4 py-2 text-xs text-gray-200 whitespace-pre font-mono">
-{entry.traceback}
-                    </pre>
-                ) : (
-                    <div className="px-4 py-3 text-xs text-gray-400">
-                        No traceback recorded for this entry.
-                    </div>
-                )}
-            </>
-        );
-    }
-    return (
-        <div className="px-4 py-3 text-xs text-gray-300 space-y-1">
-            <div>Status: <span className="font-mono">{entry.status || "n/a"}</span></div>
-            {entry.duration_ms != null && (
-                <div>Duration: <span className="font-mono">{formatDuration(entry.duration_ms)}</span></div>
-            )}
-            {entry.job_id && (
-                <div className="break-all">Job: <span className="font-mono">{entry.job_id}</span></div>
-            )}
-            <QueueRouting entry={entry}/>
-            <div className="text-gray-500 mt-2">
-                No error reported for this entry. Switch to the Metrics tab for
-                CPU / memory / IO data.
-            </div>
-        </div>
-    );
-};
-
-// Which pool a still-pending job is waiting on.
-//
-// THE ONE FACT THAT EXPLAINS A STUCK JOB, and it was not on screen anywhere. A job
-// routed to a pool no worker subscribes to is accepted and then never delivered:
-// nothing pulls it, so it is never redelivered, so it never reaches the
-// delivery-attempt cap that would record an error. The row stays `queued` with no
-// error, no retry, and no line in any worker's log — and every visible field looks
-// normal. Reading the pool off the queue entry is how that gets diagnosed in one
-// look instead of by reading queue source.
-//
-// Only for a non-terminal row: for history the pool is a spent detail, and the
-// queue entry has usually been swept anyway, which would make this a 404 on most
-// rows an operator opens.
-const QueueRouting: React.FC<{entry: AuditEntry}> = ({entry}) => {
-    const [pool, setPool] = useState<string | null | undefined>(undefined);
-    const [gone, setGone] = useState(false);
-    const status = (entry.status || "").toLowerCase();
-    const pending = !!entry.job_id && (status === "queued" || status === "running");
-
-    useEffect(() => {
-        if (!pending || !entry.job_id) return;
-        let cancelled = false;
-        (async () => {
-            try {
-                const job = await viewerApi.convertStatus(entry.job_id!);
-                if (!cancelled) setPool(job.target_capability ?? null);
-            } catch {
-                // A 404 means the entry is no longer in the queue at all, which is
-                // itself the answer for a row still claiming to be queued: there
-                // is nothing left to deliver. Not an error worth a red box.
-                if (!cancelled) setGone(true);
-            }
-        })();
-        return () => {
-            cancelled = true;
-        };
-    }, [pending, entry.job_id]);
-
-    if (!pending) return null;
-    if (gone) {
-        return (
-            <div className="text-amber-300">
-                No queue entry for this job — nothing is left to deliver, so this row will not
-                move on its own. Cancel it to clear the status.
-            </div>
-        );
-    }
-    if (pool === undefined) return <div className="text-gray-500">Pool: looking up…</div>;
-    return (
-        <div>
-            Pool: <span className="font-mono">{pool || "base"}</span>
-            {!pool && (
-                <span className="text-amber-300">
-                    {" "}— routed to the default pool, which a specialised worker does not subscribe to
-                </span>
-            )}
-        </div>
-    );
-};
-
-// Clearing a job that nothing is ever going to finish.
-//
-// Shown only for a NON-TERMINAL entry with a job id. A done or error row has
-// nothing to cancel, and putting the button there would invite an operator to
-// "fix" a row that is simply history.
-//
-// The job this is for is one queued against a capability no live worker serves
-// — a retired pool, a renamed capability, a worker that never came back. It is
-// never pulled, so it never reaches a terminal status, so the KV sweep (which
-// only touches terminal entries) never clears it: it shows as pending forever.
-// The user-facing cancel cannot reach it either, because that one filters on
-// the job's owner and an operator cleaning up after a pool is not that person.
-const StuckJobActions: React.FC<{entry: AuditEntry}> = ({entry}) => {
-    const [busy, setBusy] = useState(false);
-    const [done, setDone] = useState<string | null>(null);
-    const [err, setErr] = useState<string | null>(null);
-
-    const status = (entry.status || "").toLowerCase();
-    if (!entry.job_id || (status !== "queued" && status !== "running")) return null;
-
-    const onCancel = async () => {
-        if (
-            !confirm(
-                `Cancel job ${entry.job_id}?\n\n` +
-                "The audit row is marked cancelled and the queue entry is dropped. " +
-                "A worker that picks the message up later will see the cancellation " +
-                "and drop it. If the job is running right now it stops at its next " +
-                "cancellation check, and anything it already wrote stays written.",
-            )
-        ) {
-            return;
-        }
-        setBusy(true);
-        setErr(null);
-        try {
-            const r = await viewerApi.adminCancelJob(entry.job_id!);
-            // Both halves reported: a job can be stuck in the audit row, in the
-            // queue entry, or in both, and "cancelled" alone would leave an
-            // operator unsure whether the pending entry actually went away.
-            setDone(
-                `audit row ${r.cancelled ? "cancelled" : "unchanged"}, ` +
-                `queue entry ${r.purged ? "dropped" : "not present"}`,
-            );
-            // Reload the table the row came from. Without this the row it was
-            // cancelled from still reads `queued`, so the operator's next move is
-            // to wonder whether the cancel took and press it again -- on a job
-            // that is already gone. The store's nonce is the same mechanism the
-            // filter bar and the refresh button use, so one reload happens rather
-            // than this component fetching its own view of the world.
-            if (r.cancelled || r.purged) useAuditFilterStore.getState().refresh();
-        } catch (e) {
-            setErr(e instanceof ApiError ? e.detail || e.message : String(e));
-        } finally {
-            setBusy(false);
-        }
-    };
-
-    // Sits in the modal header, so it is laid out to fit there: a compact
-    // button, and any outcome text truncated with the whole of it on hover
-    // rather than allowed to push the close button off the row.
-    return (
-        <span className="shrink-0 flex items-center gap-2">
-            {done && (
-                <span className="text-[11px] text-gray-400 max-w-[14rem] truncate" title={done}>
-                    {done}
-                </span>
-            )}
-            {err && (
-                <span className="text-[11px] text-red-300 max-w-[14rem] truncate" title={err}>
-                    {err}
-                </span>
-            )}
-            <button
-                type="button"
-                className="text-xs bg-red-800 hover:bg-red-700 px-2 py-1 rounded-sm disabled:opacity-50 whitespace-nowrap"
-                onClick={() => void onCancel()}
-                disabled={busy || done != null}
-                title="Cancel this job and drop its queue entry (admin)"
-            >
-                {busy ? "…" : "Cancel job"}
-            </button>
-        </span>
-    );
-};
-
-// Conversion engine + effective toggles (the convert_meta JSONB). Highlights the
-// tessellator that actually ran — an "occ-builtin (fallback …)" value means adacpp
-// wasn't present in the worker so libtess2 silently degraded.
-const ConvertEngine: React.FC<{meta: import("@/services/viewerApi").ConvertMeta}> = ({meta}) => {
-    const fellBack = (meta.tessellator || "").includes("fallback");
-    const opts = meta.options || {};
-    const optKeys = Object.keys(opts).sort();
-    return (
-        <div className="pt-2 border-t border-gray-800 space-y-1">
-            <div className="text-[11px] uppercase tracking-wide text-gray-400">Conversion engine</div>
-            <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-1 font-mono">
-                {meta.tessellator && (
-                    <>
-                        <dt className="text-gray-400">Tessellator</dt>
-                        <dd className={fellBack ? "text-amber-300" : "text-emerald-300"}>{meta.tessellator}</dd>
-                    </>
-                )}
-                {meta.step_glb_pipeline && (
-                    <><dt className="text-gray-400">Requested</dt><dd>{meta.step_glb_pipeline}</dd></>
-                )}
-                {meta.glb_compression && (
-                    <><dt className="text-gray-400">Compression</dt><dd>{meta.glb_compression}</dd></>
-                )}
-                {meta.stream_workers != null && meta.stream_workers !== "" && (
-                    <><dt className="text-gray-400">Workers</dt><dd>{String(meta.stream_workers)}</dd></>
-                )}
-                {meta.convert_ms != null && (
-                    <><dt className="text-gray-400">Convert</dt><dd>{meta.convert_ms} ms</dd></>
-                )}
-                {meta.compress_ms != null && (
-                    <>
-                        <dt className="text-gray-400">Compress</dt>
-                        <dd className="text-sky-300">
-                            {meta.compress_ms} ms
-                            {meta.convert_ms != null && meta.convert_ms > 0
-                                ? ` (+${Math.round((meta.compress_ms / meta.convert_ms) * 100)}%)`
-                                : ""}
-                        </dd>
-                    </>
-                )}
-            </dl>
-            {optKeys.length > 0 && (
-                <div className="pt-1">
-                    <div className="text-[11px] text-gray-400 mb-0.5">Effective options</div>
-                    <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-0.5 font-mono text-[11px]">
-                        {optKeys.map((k) => (
-                            <React.Fragment key={k}>
-                                <dt className="text-gray-500 break-all">{k}</dt>
-                                <dd className="text-gray-300 break-all">{opts[k]}</dd>
-                            </React.Fragment>
-                        ))}
-                    </dl>
-                </div>
-            )}
-        </div>
-    );
-};
-
-// Worker package manifest ("pixi list") for the row's worker image — lazily
-// fetched on expand, filterable. Links a conversion to the exact toolchain
-// (occt / pythonocc-core / ada-cpp / …) that produced it.
-const WorkerPackages: React.FC<{imageTag: string}> = ({imageTag}) => {
-    const [open, setOpen] = useState(false);
-    const [pkgs, setPkgs] = useState<import("@/services/viewerApi").WorkerPackage[] | null>(null);
-    const [err, setErr] = useState<string | null>(null);
-    const [missing, setMissing] = useState(false);
-    const [filter, setFilter] = useState("");
-    const toggle = async () => {
-        // `missing` counts as answered: the endpoint will keep saying 404 for
-        // this worker, so re-opening must not refetch.
-        if (pkgs || err || missing) {
-            setOpen((v) => !v);
-            return;
-        }
-        setOpen(true);
-        try {
-            const r = await viewerApi.adminWorkerPackages(imageTag);
-            setPkgs(r.packages);
-        } catch (e) {
-            // Not every worker records a manifest -- see workerPackages.ts.
-            if (isMissingManifest(e)) setMissing(true);
-            else setErr(e instanceof Error ? e.message : String(e));
-        }
-    };
-    const f = filter.trim().toLowerCase();
-    const shown = (pkgs || []).filter((p) => !f || p.name.toLowerCase().includes(f));
-    return (
-        <div className="pt-2 border-t border-gray-800 space-y-1">
-            <button
-                type="button"
-                onClick={toggle}
-                className="text-[11px] text-blue-300 hover:text-blue-200 flex items-center gap-1.5"
-                title="Captured conda package manifest for this worker image"
-            >
-                <span>{open ? "▾" : "▸"}</span>
-                <span>Worker packages</span>
-                <span className="font-mono text-gray-500 break-all">{imageTag}</span>
-            </button>
-            {open && (
-                <div className="space-y-1 pl-3">
-                    {err && <div className="text-[11px] text-red-400">{err}</div>}
-                    {missing && <div className="text-[11px] text-gray-400">{MISSING_MANIFEST_NOTE}</div>}
-                    {!err && !missing && !pkgs && <div className="text-[11px] text-gray-400">Loading…</div>}
-                    {pkgs && (
-                        <>
-                            <input
-                                value={filter}
-                                onChange={(e) => setFilter(e.target.value)}
-                                placeholder="filter (occt, ada-cpp…)"
-                                className="bg-gray-900 border border-gray-700 rounded-sm px-2 py-0.5 text-[11px] text-gray-100 w-48"
-                            />
-                            <div className="max-h-64 overflow-auto">
-                                <dl className="grid grid-cols-[1fr_max-content] gap-x-3 gap-y-0.5 font-mono text-[11px]">
-                                    {shown.map((p) => (
-                                        <React.Fragment key={p.name}>
-                                            <dt className="text-gray-300 break-all">{p.name}</dt>
-                                            <dd className="text-gray-400 text-right whitespace-nowrap">
-                                                {p.version}{p.build ? ` (${p.build})` : ""}
-                                            </dd>
-                                        </React.Fragment>
-                                    ))}
-                                </dl>
-                            </div>
-                            <div className="text-[10px] text-gray-500">
-                                {shown.length} / {pkgs.length} packages
-                            </div>
-                        </>
-                    )}
-                </div>
-            )}
-        </div>
-    );
-};
-
-// adacpp [STEPPROF-JSON] pipeline summaries (captured when "Profile conversions"
-// is on): per-pipeline phase wall/RSS breakdown, kernel-exact peak (VmHWM),
-// per-solid stats, achieved parallelism / IO pressure and per-thread utilisation
-// — the C++ sibling of the Python profile below.
-const CppProfilePanel: React.FC<{profiles: import("@/services/viewerApi").CppProfile[]}> = ({profiles}) => (
-    <div className="pt-2 border-t border-gray-800 space-y-3">
-        <div className="text-[11px] uppercase tracking-wide text-gray-400">C++ pipeline profile</div>
-        {profiles.map((p, i) => {
-            const wall = p.wall_ms > 0 ? p.wall_ms : 1;
-            return (
-                <div key={`${p.label}-${i}`} className="space-y-1">
-                    <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-0.5 font-mono text-[11px]">
-                        <dt className="text-gray-400">Pipeline</dt>
-                        <dd className="text-emerald-300">{p.label}</dd>
-                        <dt className="text-gray-400">Wall</dt>
-                        <dd>{formatDuration(p.wall_ms)}</dd>
-                        <dt className="text-gray-400">Peak RSS</dt>
-                        <dd>{formatBytes(p.peak_rss_mb * 1024 * 1024)} <span className="text-gray-500">(VmHWM)</span></dd>
-                        {p.cpu_s != null && (
-                            <><dt className="text-gray-400">CPU</dt>
-                            <dd>{p.cpu_s.toFixed(1)} s{p.parallelism != null ? ` — ${p.parallelism.toFixed(2)}x cores busy` : ""}</dd></>
-                        )}
-                        {(p.solids ?? 0) > 0 && (
-                            <><dt className="text-gray-400">Solids</dt>
-                            <dd>{p.solids}{(p.tris ?? 0) > 0 ? ` (${p.tris} tris, max ${p.max_tris_solid}/solid)` : ""}</dd></>
-                        )}
-                        {p.disk_read_mb != null && p.disk_read_mb > 0 && (
-                            <><dt className="text-gray-400">Disk read</dt>
-                            <dd>{p.disk_read_mb.toFixed(0)} MB physical{p.majflt != null ? `, ${p.majflt} major faults` : ""}</dd></>
-                        )}
-                    </dl>
-                    {p.phases.length > 0 && (
-                        <table className="w-full text-[11px] font-mono">
-                            <thead>
-                                <tr className="text-gray-500 text-left">
-                                    <th className="font-normal pr-2">phase</th>
-                                    <th className="font-normal pr-2 text-right">ms</th>
-                                    <th className="font-normal pr-2 text-right">RSS</th>
-                                    <th className="font-normal w-1/3">share</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {p.phases.map((ph) => (
-                                    <tr key={ph.name} className="text-gray-300">
-                                        <td className="pr-2 break-all">{ph.name}</td>
-                                        <td className="pr-2 text-right">{Math.round(ph.ms)}</td>
-                                        <td className="pr-2 text-right">{Math.round(ph.rss_mb)} MB</td>
-                                        <td>
-                                            <div className="bg-gray-800 rounded-sm h-2 w-full">
-                                                <div
-                                                    className="bg-sky-600 rounded-sm h-2"
-                                                    style={{width: `${Math.min(100, (ph.ms / wall) * 100).toFixed(1)}%`}}
-                                                />
-                                            </div>
-                                        </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
-                    )}
-                    {p.notes && Object.keys(p.notes).length > 0 && (
-                        <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-0.5 font-mono text-[11px]">
-                            {Object.entries(p.notes).map(([k, v]) => (
-                                <React.Fragment key={k}>
-                                    <dt className="text-gray-500 break-all">{k}</dt>
-                                    <dd className="text-gray-300">{v}</dd>
-                                </React.Fragment>
-                            ))}
-                        </dl>
-                    )}
-                    {(p.threads?.length ?? 0) > 0 && (
-                        <div className="text-[11px] font-mono text-gray-400">
-                            threads:{" "}
-                            {p.threads!.map((t) => `t${t.tid}=${Math.round(t.busy_ms)}ms/${t.solids}s`).join("  ")}
-                        </div>
-                    )}
-                </div>
-            );
-        })}
-    </div>
-);
-
-const MetricsTab: React.FC<{
-    entry: AuditEntry;
-    onDownloadProfile: () => void;
-    downloading: boolean;
-    downloadErr: string | null;
-}> = ({entry, onDownloadProfile, downloading, downloadErr}) => {
-    if (!hasMetrics(entry)) {
-        return (
-            <div className="px-4 py-3 text-xs text-gray-400">
-                No metrics captured for this entry.
-                {entry.action !== "convert" ? (
-                    <span> Only conversion runs collect resource metrics; a compile run records
-                        its wall time and its engine log.</span>
-                ) : (
-                    <span> The worker that processed this job pre-dates the metrics column.</span>
-                )}
-            </div>
-        );
-    }
-    return (
-        <div className="px-4 py-3 text-xs text-gray-200 space-y-3">
-            <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-1 font-mono">
-                <MetricRow label="Wall" value={formatDuration(entry.duration_ms)}/>
-                <MetricRow label="CPU user" value={formatDuration(entry.cpu_user_ms)}/>
-                <MetricRow label="CPU sys"  value={formatDuration(entry.cpu_sys_ms)}/>
-                <MetricRow label="Peak RSS" value={formatBytes((entry.peak_rss_kb ?? null) != null ? (entry.peak_rss_kb as number) * 1024 : null)}/>
-                <MetricRow label="Read"     value={formatBytes(entry.read_bytes)}/>
-                <MetricRow label="Write"    value={formatBytes(entry.write_bytes)}/>
-            </dl>
-            {entry.convert_meta && <ConvertEngine meta={entry.convert_meta}/>}
-            {(entry.convert_meta?.cpp_profile?.length ?? 0) > 0 && (
-                <CppProfilePanel profiles={entry.convert_meta!.cpp_profile!}/>
-            )}
-            {entry.worker_image_tag && <WorkerPackages imageTag={entry.worker_image_tag}/>}
-            <MetricsHistoryChart
-                auditId={entry.id}
-                cores={entry.convert_meta?.cpu_cores ?? undefined}
-                native={entry.convert_meta?.tessellator === "adacpp:native"}
-            />
-            {entry.profile_key ? (
-                <div className="pt-2 border-t border-gray-800 space-y-3">
-                    <div>
-                        <button
-                            type="button"
-                            className="bg-blue-700 hover:bg-blue-600 px-3 py-1 rounded-sm text-xs disabled:opacity-50"
-                            onClick={onDownloadProfile}
-                            disabled={downloading}
-                        >
-                            {downloading ? "Downloading…" : "Download profile (.prof)"}
-                        </button>
-                        <span className="text-[10px] text-gray-500 ml-2">
-                            Loadable in snakeviz / speedscope / pstats.
-                        </span>
-                        {downloadErr && (
-                            <div className="text-red-300 text-[10px] mt-1 break-all">{downloadErr}</div>
-                        )}
-                    </div>
-                    {/* Inline per-function stats — sortable and searchable so
-                        the operator can find hot frames without leaving the
-                        UI. Server-side pstats parse keeps the SPA bundle
-                        free of marshal/pickle parsers. */}
-                    <ProfileStatsTable auditId={entry.id} totalWallMs={entry.duration_ms}/>
-                </div>
-            ) : (
-                <div className="text-[10px] text-gray-500 pt-2 border-t border-gray-800">
-                    No profile attached. Toggle "Profile conversions" above and re-run.
-                </div>
-            )}
-        </div>
-    );
-};
-
-// Per-row inspection of a browser view/render event: the full
-// client_metrics payload (lazily fetched), shown as a phase/scalar grid
-// plus the per-function self-time frames when call profiling was on.
-const CM_MS_KEYS = new Set([
-    "total_ms", "ttfb_ms", "download_ms", "decompress_ms", "parse_ms", "prepare_ms",
-    "first_render_ms", "dns_ms", "tcp_ms", "tls_ms", "long_task_ms", "blocking_ms",
-    "window_ms", "frame_ms_p50", "frame_ms_p95", "gpu_ms_p50", "gpu_ms_p95",
-]);
-const CM_BYTE_KEYS = new Set(["transfer_bytes", "decoded_bytes"]);
-
-const ClientMetricsTab: React.FC<{entry: AuditEntry}> = ({entry}) => {
-    const [cm, setCm] = useState<Record<string, unknown> | null>(null);
-    const [err, setErr] = useState<string | null>(null);
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        let alive = true;
-        setLoading(true);
-        viewerApi
-            .adminAuditClientMetrics(entry.id)
-            .then((d) => alive && setCm(d.client_metrics))
-            .catch((e) => alive && setErr((e as Error).message))
-            .finally(() => alive && setLoading(false));
-        return () => {
-            alive = false;
-        };
-    }, [entry.id]);
-
-    if (loading) return <div className="px-4 py-3 text-xs text-gray-400">Loading…</div>;
-    if (err) return <div className="px-4 py-3 text-xs text-red-300">{err}</div>;
-    if (!cm) return <div className="px-4 py-3 text-xs text-gray-400">No client metrics recorded for this entry.</div>;
-
-    const frames = Array.isArray(cm.profile_frames)
-        ? (cm.profile_frames as Array<{fn?: string; self_ms?: number; total_ms?: number}>)
-        : [];
-    // Scalar fields → a sorted key/value grid, with ms/bytes formatting.
-    const scalars = Object.entries(cm)
-        .filter(([k, v]) => k !== "profile_frames" && (typeof v === "number" || typeof v === "string" || typeof v === "boolean"))
-        .sort(([a], [b]) => a.localeCompare(b));
-    // Performance toggles active when this load/render was captured.
-    const perfOptions =
-        cm.perf_options && typeof cm.perf_options === "object" && !Array.isArray(cm.perf_options)
-            ? (cm.perf_options as Record<string, unknown>)
-            : null;
-    const fmt = (k: string, v: unknown): string => {
-        if (typeof v === "number") {
-            if (CM_MS_KEYS.has(k)) return formatDuration(v);
-            if (CM_BYTE_KEYS.has(k)) return formatBytes(v);
-            return String(Math.round(v * 100) / 100);
-        }
-        return String(v);
-    };
-
-    return (
-        <div className="px-4 py-3 text-xs text-gray-200 space-y-3">
-            <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-0.5 font-mono">
-                {scalars.map(([k, v]) => (
-                    <React.Fragment key={k}>
-                        <dt className="text-gray-400">{k}</dt>
-                        <dd className="text-gray-100">{fmt(k, v)}</dd>
-                    </React.Fragment>
-                ))}
-            </dl>
-            {perfOptions ? (
-                <div className="pt-2 border-t border-gray-800">
-                    <div className="text-[11px] text-gray-400 mb-1">Performance options (active at capture)</div>
-                    <dl className="grid grid-cols-[max-content_1fr] gap-x-4 gap-y-0.5 font-mono">
-                        {Object.entries(perfOptions)
-                            .sort(([a], [b]) => a.localeCompare(b))
-                            .map(([k, v]) => (
-                                <React.Fragment key={k}>
-                                    <dt className="text-gray-400">{k}</dt>
-                                    <dd className="text-gray-100">{String(v)}</dd>
-                                </React.Fragment>
-                            ))}
-                    </dl>
-                </div>
-            ) : null}
-            {frames.length > 0 ? (
-                <div className="pt-2 border-t border-gray-800">
-                    <div className="text-[11px] text-gray-400 mb-1">
-                        Call hotspots (self-time, TS + WASM)
-                        {entry.action === "render" && " · main-thread only (GPU-bound shows in gpu_ms)"}
-                    </div>
-                    <table className="w-full">
-                        <thead className="text-gray-400">
-                            <tr>
-                                <th className="text-left font-medium">Function</th>
-                                <th className="text-right font-medium">Self</th>
-                                <th className="text-right font-medium">Total</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {frames.map((f, i) => (
-                                <tr key={i} className="border-t border-gray-800">
-                                    <td className="py-0.5 font-mono truncate max-w-md" title={f.fn}>
-                                        {typeof f.fn === "string" && f.fn.toLowerCase().includes("wasm") && (
-                                            <span className="text-violet-300 mr-1">[wasm]</span>
-                                        )}
-                                        {f.fn}
-                                    </td>
-                                    <td className="text-right font-mono">{formatDuration(f.self_ms ?? null)}</td>
-                                    <td className="text-right font-mono">{formatDuration(f.total_ms ?? null)}</td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
-                </div>
-            ) : (
-                <div className="text-[10px] text-gray-500 pt-2 border-t border-gray-800">
-                    No call profile on this event. Enable "Profile calls" in Performance options
-                    (Chromium + Document-Policy: js-profiling).
-                </div>
-            )}
-        </div>
-    );
-};
-
-const MetricsHistoryChart: React.FC<{auditId: number; cores?: number; native?: boolean}> = ({
-    auditId,
-    cores,
-    native,
-}) => {
-    const [samples, setSamples] = useState<MetricsSample[] | null>(null);
-    const [err, setErr] = useState<string | null>(null);
-    const [loading, setLoading] = useState(false);
-
-    useEffect(() => {
-        let cancelled = false;
-        setLoading(true);
-        setErr(null);
-        viewerApi.adminMetricsHistory(auditId)
-            .then((r) => {
-                if (!cancelled) setSamples(r.samples);
-            })
-            .catch((e) => {
-                if (!cancelled) setErr(e instanceof ApiError ? e.detail || e.message : String(e));
-            })
-            .finally(() => {
-                if (!cancelled) setLoading(false);
-            });
-        return () => {
-            cancelled = true;
-        };
-    }, [auditId]);
-
-    if (loading) {
-        return <div className="text-[10px] text-gray-400">Loading resource timeline…</div>;
-    }
-    if (err) {
-        return <div className="text-[10px] text-red-300 break-all">timeline: {err}</div>;
-    }
-    if (!samples || samples.length === 0) {
-        return (
-            <div className="text-[10px] text-gray-500 pt-1 border-t border-gray-800">
-                No per-heartbeat samples for this run. Older jobs predate the
-                subprocess wrapper, or the worker died before any sample landed.
-            </div>
-        );
-    }
-
-    const maxElapsed = Math.max(...samples.map((s) => s.elapsed_s), 1);
-    const maxRss = Math.max(...samples.map((s) => Math.max(s.rss_kb, s.peak_rss_kb)), 1);
-    const maxIo = Math.max(...samples.map((s) => Math.max(s.read_bytes, s.write_bytes)), 1);
-
-    // cpu_user_ms/cpu_sys_ms are CUMULATIVE, so plotting them is just a monotonic ramp. Instead show
-    // utilization: Δ(user+sys) / Δwall between consecutive samples = CPU cores busy in that window.
-    // When the pod's core count is known (convert_meta.cpu_cores) render it as % of all cores (0–100%);
-    // otherwise as absolute cores busy.
-    const cpuCoresBusy: number[] = samples.map((s, i) => {
-        const prevCpu = i === 0 ? 0 : samples[i - 1].cpu_user_ms + samples[i - 1].cpu_sys_ms;
-        const prevT = i === 0 ? 0 : samples[i - 1].elapsed_s;
-        const dCpuMs = s.cpu_user_ms + s.cpu_sys_ms - prevCpu;
-        const dWallMs = (s.elapsed_s - prevT) * 1000;
-        return dWallMs > 0 ? Math.max(0, dCpuMs / dWallMs) : 0;
-    });
-    const maxCoresBusy = Math.max(...cpuCoresBusy, 0.01);
-    const cpuFullScale = cores && cores > 0 ? cores : Math.max(Math.ceil(maxCoresBusy), 1);
-    const cpuLabel =
-        cores && cores > 0
-            ? `${Math.round((maxCoresBusy / cores) * 100)}% of ${cores} cores`
-            : `${maxCoresBusy.toFixed(1)} cores`;
-
-    // Per-thread (≈ per-core) utilization envelope — only for the in-process native engine, where the
-    // C++ tessellation threads live in the sampled process. Match threads by tid across consecutive
-    // samples, take each thread's Δcpu/Δwall (1.0 = one core's worth), then SORT descending per sample
-    // so each plotted line is a "core rank": N lines pinned high = N cores saturated, vs all N at
-    // mid-height = N cores half-busy — the distinction the aggregate % can't make.
-    const hasPerThread =
-        !!native && samples.some((s) => s.per_thread_cpu_ms && Object.keys(s.per_thread_cpu_ms).length > 0);
-    const nLanes = hasPerThread ? Math.max(cores && cores > 0 ? cores : 1, 1) : 0;
-    const lanes: number[][] = hasPerThread
-        ? samples.map((s, i) => {
-              const cur = s.per_thread_cpu_ms || {};
-              const prev = i > 0 ? samples[i - 1].per_thread_cpu_ms || {} : {};
-              const dWallMs = (s.elapsed_s - (i > 0 ? samples[i - 1].elapsed_s : 0)) * 1000;
-              const busy = Object.keys(cur)
-                  .map((tid) => (dWallMs > 0 ? Math.max(0, (cur[tid] - (prev[tid] ?? 0)) / dWallMs) : 0))
-                  .sort((a, b) => b - a);
-              return Array.from({length: nLanes}, (_, k) => busy[k] ?? 0); // rank-ordered, padded to nLanes
-          })
-        : [];
-    const laneColors = Array.from({length: nLanes}, (_, k) => `hsl(152 ${72 - k * 6}% ${62 - k * 5}%)`);
-    const peakLanes = hasPerThread
-        ? Math.max(0, ...samples.map((_, i) => lanes[i].filter((v) => v >= 0.5).length))
-        : 0;
-
-    return (
-        <div className="space-y-2 pt-2 border-t border-gray-800">
-            <div className="text-[10px] uppercase tracking-wide text-gray-500">
-                Resource timeline ({samples.length} samples · {maxElapsed.toFixed(0)}s)
-            </div>
-            <ChartPanel
-                title="RSS"
-                yLabel={formatBytes(maxRss * 1024)}
-                series={[
-                    {color: "#60a5fa", points: samples.map((s) => [s.elapsed_s / maxElapsed, s.rss_kb / maxRss]), label: "rss"},
-                    {color: "#f87171", points: samples.map((s) => [s.elapsed_s / maxElapsed, s.peak_rss_kb / maxRss]), label: "peak"},
-                ]}
-            />
-            {hasPerThread ? (
-                <CpuPerCorePanel
-                    samples={samples}
-                    lanes={lanes}
-                    nLanes={nLanes}
-                    cores={cores && cores > 0 ? cores : nLanes}
-                    maxElapsed={maxElapsed}
-                    colors={laneColors}
-                    peakLanes={peakLanes}
-                />
-            ) : (
-                <ChartPanel
-                    title={cores ? "CPU util (% of cores)" : "CPU (cores busy)"}
-                    yLabel={cpuLabel}
-                    series={[
-                        {
-                            color: "#34d399",
-                            points: samples.map((s, i) => [s.elapsed_s / maxElapsed, Math.min(1, cpuCoresBusy[i] / cpuFullScale)]),
-                            label: "cpu",
-                        },
-                    ]}
-                />
-            )}
-            <ChartPanel
-                title="IO bytes"
-                yLabel={formatBytes(maxIo)}
-                series={[
-                    {color: "#a78bfa", points: samples.map((s) => [s.elapsed_s / maxElapsed, s.read_bytes / maxIo]), label: "read"},
-                    {color: "#fbbf24", points: samples.map((s) => [s.elapsed_s / maxElapsed, s.write_bytes / maxIo]), label: "write"},
-                ]}
-            />
-        </div>
-    );
-};
-
-// Stacked per-core CPU utilization (native engine). Each band is one core, rank-sorted busiest →
-// least and stacked from the bottom; the y-axis ceiling is the pod's core count. So the FILLED
-// HEIGHT at any instant = "cores busy," and saturation reads directly off the chart: the stack
-// hugging the top across the whole width = all cores saturated for the whole conversion; a dip = an
-// idle stretch. 3 full bands reaching 3/4 (native reserves one core for the parent) is unmistakably
-// different from 4 half-height bands reaching 2/4 — which the aggregate % and overlaid lines blur.
-const CpuPerCorePanel: React.FC<{
-    samples: MetricsSample[];
-    lanes: number[][];
-    nLanes: number;
-    cores: number;
-    maxElapsed: number;
-    colors: string[];
-    peakLanes: number;
-}> = ({samples, lanes, nLanes, cores, maxElapsed, colors, peakLanes}) => {
-    const W = 320;
-    const H = 56;
-    const x = (i: number) => (samples[i].elapsed_s / maxElapsed) * W;
-    const yAt = (cumCores: number) => H - Math.min(1, cumCores / cores) * H;
-    const cum = (i: number, k: number) => {
-        let c = 0;
-        for (let j = 0; j < k; j++) c += lanes[i][j] ?? 0;
-        return c;
-    };
-    return (
-        <div>
-            <div className="flex items-baseline justify-between gap-2 mb-0.5">
-                <div className="text-[10px] text-gray-400 font-mono">CPU per core (native)</div>
-                <div className="text-[10px] text-gray-500 font-mono">
-                    top = {cores} cores · peak {peakLanes} saturated
-                </div>
-            </div>
-            <svg viewBox={`0 0 ${W} ${H}`} className="w-full h-14 bg-gray-900/60 border border-gray-800 rounded-sm">
-                {Array.from({length: nLanes}, (_, k) => {
-                    const top = samples.map(
-                        (_, i) => `${i === 0 ? "M" : "L"}${x(i).toFixed(1)},${yAt(cum(i, k + 1)).toFixed(1)}`,
-                    );
-                    const bot: string[] = [];
-                    for (let i = samples.length - 1; i >= 0; i--) {
-                        bot.push(`L${x(i).toFixed(1)},${yAt(cum(i, k)).toFixed(1)}`);
-                    }
-                    return (
-                        <path key={k} d={`${top.join(" ")} ${bot.join(" ")} Z`} fill={colors[k]} fillOpacity={0.85} stroke="none"/>
-                    );
-                })}
-            </svg>
-        </div>
-    );
-};
-
-const ChartPanel: React.FC<{
-    title: string;
-    yLabel: string;
-    series: {color: string; points: [number, number][]; label: string}[];
-}> = ({title, yLabel, series}) => {
-    const W = 320;
-    const H = 56;
-    return (
-        <div>
-            <div className="flex items-baseline justify-between gap-2 mb-0.5">
-                <div className="text-[10px] text-gray-400 font-mono">{title}</div>
-                <div className="text-[10px] text-gray-500 font-mono">max ≈ {yLabel}</div>
-            </div>
-            <svg viewBox={`0 0 ${W} ${H}`} className="w-full h-14 bg-gray-900/60 border border-gray-800 rounded-sm">
-                {series.map((s, i) => {
-                    if (s.points.length === 0) return null;
-                    const d = s.points
-                        .map(([x, y], idx) =>
-                            `${idx === 0 ? "M" : "L"}${(x * W).toFixed(2)},${(H - y * H).toFixed(2)}`,
-                        )
-                        .join(" ");
-                    return (
-                        <g key={i}>
-                            <path d={d} stroke={s.color} strokeWidth={1.2} fill="none"/>
-                        </g>
-                    );
-                })}
-            </svg>
-            <div className="flex gap-3 mt-0.5 text-[10px] text-gray-500 font-mono">
-                {series.map((s, i) => (
-                    <span key={i} style={{color: s.color}}>● <span className="text-gray-400">{s.label}</span></span>
-                ))}
-            </div>
-        </div>
-    );
-};
-
-
-type StatsSortKey = "func" | "ncalls" | "primitive_calls" | "tottime" | "percall_tot" | "cumtime" | "percall_cum";
-
-const ProfileStatsTable: React.FC<{auditId: number; totalWallMs: number | null}> = ({auditId, totalWallMs}) => {
-    const [resp, setResp] = useState<{rows: ProfileStatsRow[]; total_tottime: number} | null>(null);
-    const [loading, setLoading] = useState(false);
-    const [err, setErr] = useState<string | null>(null);
-    const [filter, setFilter] = useState("");
-    const [sortKey, setSortKey] = useState<StatsSortKey>("cumtime");
-    const [sortDesc, setSortDesc] = useState(true);
-
-    useEffect(() => {
-        let cancelled = false;
-        setLoading(true);
-        setErr(null);
-        viewerApi.adminProfileStats(auditId, 1000)
-            .then((r) => {
-                if (cancelled) return;
-                setResp({rows: r.rows, total_tottime: r.total_tottime});
-            })
-            .catch((e) => {
-                if (cancelled) return;
-                setErr(e instanceof ApiError ? e.detail || e.message : String(e));
-            })
-            .finally(() => {
-                if (cancelled) return;
-                setLoading(false);
-            });
-        return () => {
-            cancelled = true;
-        };
-    }, [auditId]);
-
-    const filtered = useMemo(() => {
-        if (!resp) return [];
-        const q = filter.trim().toLowerCase();
-        const rows = q
-            ? resp.rows.filter(
-                  (r) =>
-                      r.func.toLowerCase().includes(q) ||
-                      r.file.toLowerCase().includes(q),
-              )
-            : resp.rows;
-        const sorted = [...rows];
-        sorted.sort((a, b) => {
-            const av = a[sortKey];
-            const bv = b[sortKey];
-            if (typeof av === "number" && typeof bv === "number") return sortDesc ? bv - av : av - bv;
-            return sortDesc ? String(bv).localeCompare(String(av)) : String(av).localeCompare(String(bv));
-        });
-        return sorted;
-    }, [resp, filter, sortKey, sortDesc]);
-
-    if (loading) {
-        return (
-            <div className="text-[10px] text-gray-400">Loading profile stats…</div>
-        );
-    }
-    if (err) {
-        return (
-            <div className="text-[10px] text-red-300 break-all">profile stats: {err}</div>
-        );
-    }
-    if (!resp || resp.rows.length === 0) {
-        return <div className="text-[10px] text-gray-400">No frames recorded.</div>;
-    }
-
-    const onHeader = (k: StatsSortKey) => {
-        if (k === sortKey) {
-            setSortDesc((v) => !v);
-        } else {
-            setSortKey(k);
-            setSortDesc(k !== "func");
-        }
-    };
-
-    // Top-N visual: a single horizontal bar per function (cumtime as % of
-    // total tottime) gives a "where did the time go" cue alongside the
-    // table — same Pareto graph snakeviz draws, just inline.
-    const topByCum = [...resp.rows].sort((a, b) => b.cumtime - a.cumtime).slice(0, 10);
-    const maxCum = topByCum.length > 0 ? topByCum[0].cumtime : 1;
-
-    return (
-        <div className="text-[11px] space-y-2">
-            <div className="text-gray-400">
-                <span className="font-mono">{resp.rows.length}</span> functions ·{" "}
-                <span className="font-mono">{resp.total_tottime.toFixed(2)}s</span> total self-time
-                {totalWallMs != null ? (
-                    <span> · wall <span className="font-mono">{(totalWallMs / 1000).toFixed(2)}s</span></span>
-                ) : null}
-            </div>
-            {/* Top-10 cumulative-time bar chart */}
-            <div className="space-y-0.5 bg-gray-900/60 border border-gray-800 rounded-sm p-2">
-                <div className="text-[10px] uppercase tracking-wide text-gray-500 mb-1">
-                    Top 10 by cumulative time
-                </div>
-                {topByCum.map((r, i) => (
-                    <div key={`${r.file}:${r.line}:${r.func}:${i}`} className="flex items-center gap-2">
-                        <div className="flex-1 min-w-0">
-                            <div className="font-mono text-gray-200 truncate" title={`${r.file}:${r.line}`}>
-                                {r.func}
-                            </div>
-                            <div className="h-1 bg-gray-800 rounded-sm overflow-hidden">
-                                <div
-                                    className="h-full bg-blue-500"
-                                    style={{width: `${maxCum > 0 ? (r.cumtime / maxCum) * 100 : 0}%`}}
-                                />
-                            </div>
-                        </div>
-                        <div className="font-mono text-gray-300 w-16 text-right shrink-0">
-                            {r.cumtime.toFixed(3)}s
-                        </div>
-                    </div>
-                ))}
-            </div>
-            <input
-                type="text"
-                value={filter}
-                onChange={(e) => setFilter(e.target.value)}
-                placeholder="Filter by function or file…"
-                className="bg-gray-800 border border-gray-700 rounded-sm px-2 py-1 w-full text-white"
-            />
-            <div className="overflow-auto max-h-96 border border-gray-800 rounded-sm">
-                <table className="w-full font-mono">
-                    <thead className="sticky top-0 bg-gray-800 text-gray-300">
-                        <tr className="text-right">
-                            <StatsTh active={sortKey === "func"} desc={sortDesc} onClick={() => onHeader("func")} align="left">function</StatsTh>
-                            <StatsTh active={sortKey === "ncalls"} desc={sortDesc} onClick={() => onHeader("ncalls")}>ncalls</StatsTh>
-                            <StatsTh active={sortKey === "primitive_calls"} desc={sortDesc} onClick={() => onHeader("primitive_calls")}>prim</StatsTh>
-                            <StatsTh active={sortKey === "tottime"} desc={sortDesc} onClick={() => onHeader("tottime")}>tottime</StatsTh>
-                            <StatsTh active={sortKey === "percall_tot"} desc={sortDesc} onClick={() => onHeader("percall_tot")}>percall</StatsTh>
-                            <StatsTh active={sortKey === "cumtime"} desc={sortDesc} onClick={() => onHeader("cumtime")}>cumtime</StatsTh>
-                            <StatsTh active={sortKey === "percall_cum"} desc={sortDesc} onClick={() => onHeader("percall_cum")}>percum</StatsTh>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {filtered.map((r, i) => (
-                            <tr key={`${r.file}:${r.line}:${r.func}:${i}`} className="border-t border-gray-800 text-right">
-                                <td className="px-2 py-0.5 text-left text-gray-200 truncate max-w-[20rem]" title={`${r.file}:${r.line}`}>
-                                    <span className="text-gray-500">{shortFile(r.file)}:{r.line} </span>
-                                    {r.func}
-                                </td>
-                                <td className="px-2 py-0.5 text-gray-300">{r.ncalls.toLocaleString()}</td>
-                                <td className="px-2 py-0.5 text-gray-400">{r.primitive_calls.toLocaleString()}</td>
-                                <td className="px-2 py-0.5 text-gray-200">{r.tottime.toFixed(3)}</td>
-                                <td className="px-2 py-0.5 text-gray-400">{formatPercall(r.percall_tot)}</td>
-                                <td className="px-2 py-0.5 text-gray-200">{r.cumtime.toFixed(3)}</td>
-                                <td className="px-2 py-0.5 text-gray-400">{formatPercall(r.percall_cum)}</td>
-                            </tr>
-                        ))}
-                    </tbody>
-                </table>
-                {filtered.length === 0 && (
-                    <div className="text-center text-gray-500 py-3">No frames match the filter.</div>
-                )}
-            </div>
-        </div>
-    );
-};
-
-const StatsTh: React.FC<{
-    active: boolean;
-    desc: boolean;
-    onClick: () => void;
-    align?: "left" | "right";
-    children: React.ReactNode;
-}> = ({active, desc, onClick, align = "right", children}) => (
-    <th
-        className={
-            "px-2 py-1 select-none cursor-pointer hover:text-white whitespace-nowrap " +
-            (align === "left" ? "text-left" : "text-right")
-        }
-        onClick={onClick}
-    >
-        {children}
-        {active ? <span className="text-blue-400">{desc ? " ▾" : " ▴"}</span> : null}
-    </th>
-);
-
-function shortFile(p: string): string {
-    if (!p) return "";
-    const parts = p.split("/");
-    if (parts.length <= 2) return p;
-    return ".../" + parts.slice(-2).join("/");
-}
-
-function formatPercall(v: number): string {
-    if (!isFinite(v) || v === 0) return "0";
-    if (v >= 1) return v.toFixed(3);
-    if (v >= 1e-3) return (v * 1000).toFixed(2) + "ms";
-    return (v * 1e6).toFixed(1) + "µs";
-}
-
-const MetricRow: React.FC<{label: string; value: string}> = ({label, value}) => (
-    <>
-        <dt className="text-gray-400">{label}</dt>
-        <dd>{value}</dd>
-    </>
-);
-
-// The audit log keeps its own conventions: IEC unit labels, an en dash for a
-// missing value, and durations precise enough to compare two runs of one cell.
-const formatDuration = (ms: number | null) => formatMillis(ms, {empty: "–", precise: true});
-const formatBytes = (n: number | null) => formatBytesBase(n, {units: "iec", empty: "–"});
-
-const Th: React.FC<{children: React.ReactNode}> = ({children}) => (
-    <th className="px-3 py-2 font-medium text-gray-300 whitespace-nowrap">{children}</th>
-);
-
-// Truncation lives at the cell level so long values (paths, error
-// messages, full subs) don't break layout — but we let the column
-// widths do the gating now via <colgroup>, not a hard 20ch cap.
-const Td: React.FC<{children: React.ReactNode; title?: string}> = ({children, title}) => (
-    <td className="px-3 py-1 truncate" title={title}>
-        {children}
-    </td>
-);
-
-// In-browser (WASM) conversions get a ``wasm-`` job id and a ``wasm:``
-// image tag from the audit/local endpoints; surface either as a badge so
-// operators can tell client-side rows from worker rows at a glance.
-function isWasmEntry(e: {job_id?: string | null; worker_image_tag?: string | null}): boolean {
-    return (e.job_id || "").startsWith("wasm-") || (e.worker_image_tag || "").startsWith("wasm:");
-}
-
-const WasmBadge: React.FC = () => (
-    <span
-        className="ml-1 inline-block rounded bg-purple-900/60 text-purple-200 px-1 text-[9px] font-semibold align-middle"
-        title="Converted in-browser (WASM)"
-    >
-        WASM
-    </span>
-);
-
-function shortSub(s: string | null): string {
-    if (!s) return "";
-    if (s.length <= 12) return s;
-    return `${s.slice(0, 8)}…${s.slice(-4)}`;
-}
-
-// Hover tooltip for the user column: display name + email (resolved server-side from the users
-// table), falling back to the raw subject when a name/email isn't on file.
-function userTooltip(e: AuditEntry): string {
-    const lines: string[] = [];
-    if (e.user_display_name) lines.push(e.user_display_name);
-    if (e.user_email) lines.push(e.user_email);
-    if (e.user_sub) lines.push(e.user_sub);
-    return lines.join("\n");
-}
-
-function formatTs(ts: string | null): string {
-    // Local time, not the raw UTC string: an earlier version sliced the wire
-    // value verbatim and the table read hours behind for any viewer east of UTC.
-    return localDateTime(ts);
-}
-
-
-// Per-row badge for the issue-bot's sync status on failed user
-// conversions (M5b). Only renders on rows the bot is actually
-// allowed to touch: failed conversions not attached to an audit
-// run (audit-run-attached rows go through the parent run's bot
-// pass — showing a badge here would imply double-processing). A
-// row that hasn't been claimed yet (status NULL) shows a
-// 'pending' pill so the operator knows the bot will get to it.
-const ISSUE_BOT_BADGE_LOG: Record<string, {cls: string; label: string}> = {
-    done:    {cls: "bg-emerald-900/40 border-emerald-700 text-emerald-200", label: "issue synced"},
-    skipped: {cls: "bg-gray-800 border-gray-600 text-gray-400",             label: "bot disabled"},
-    failed:  {cls: "bg-red-900/40 border-red-700 text-red-200",             label: "issue sync failed"},
-    syncing: {cls: "bg-blue-900/40 border-blue-700 text-blue-200",          label: "syncing…"},
-};
-
-const IssueBotBadge: React.FC<{
-    entry: AuditEntry;
-    onChanged: () => void;
-}> = ({entry, onChanged}) => {
-    const [busy, setBusy] = useState(false);
-    const [err, setErr] = useState<string | null>(null);
-    const isFailed = entry.status === "error" || entry.status === "failed";
-    if (!isFailed) return null;
-    if (entry.audit_run_id) {
-        // Audit-run-attached rows: deliberately silent — the parent
-        // run's badge tracks the bot for the batch.
-        return null;
-    }
-    const status = entry.issue_bot_status;
-    const badge = status
-        ? ISSUE_BOT_BADGE_LOG[status] || {
-            cls: "bg-gray-800 border-gray-600 text-gray-400",
-            label: status,
-        }
-        : {cls: "bg-amber-900/30 border-amber-700 text-amber-300", label: "issue pending"};
-    const retry = async () => {
-        setBusy(true);
-        setErr(null);
-        try {
-            await viewerApi.adminAuditLogSyncIssue(entry.id);
-            // Give the background task a moment to flip status before
-            // we re-fetch — otherwise the badge briefly snaps back
-            // to its old value.
-            await new Promise((r) => setTimeout(r, 600));
-            onChanged();
-        } catch (e) {
-            setErr((e as Error).message || "retry failed");
-        } finally {
-            setBusy(false);
-        }
-    };
-    return (
-        <span className="ml-2 inline-flex items-center gap-1">
-            <span
-                className={`px-1.5 py-0.5 rounded-sm border text-[10px] ${badge.cls}`}
-                title={entry.issue_bot_last_error || badge.label}
-            >
-                {badge.label}
-            </span>
-            {(status === "failed" || status === "done") && (
-                <button
-                    type="button"
-                    onClick={retry}
-                    disabled={busy}
-                    className="text-[10px] text-blue-400 hover:text-blue-300 disabled:opacity-50 no-drag"
-                    title="Re-run the issue-bot for this row"
-                >
-                    {busy ? "queued…" : "resync"}
-                </button>
-            )}
-            {err && (
-                <span className="text-[10px] text-red-400" role="alert">{err}</span>
-            )}
-        </span>
-    );
-};
-
-function statusClass(s: string | null): string {
-    if (s === "ok" || s === "done") return "text-green-400";
-    if (s === "error") return "text-red-400";
-    if (s === "queued") return "text-yellow-300";
-    return "text-gray-300";
-}
-
-export default AuditLogTab;
